Fixed a bunch of bugs in the cross validation script and added new features.

+++ code/crossvalidate.pl
@@ -23,7 +23,7 @@ use Getopt::Std;
 use POSIX;
 
 local $| = 1;
-our ($opt_k,$opt_p,$opt_e);
+our ($opt_k,$opt_p,$opt_e,$opt_i,$opt_v,$opt_d);
 my $usage = "\nK-Fold Cross-Validater\n\n" . 
 			"Usage: ./crossvalidate.pl [OPTIONS] SEGMENTATION_COMMAND TRUE_CORPUS\n\n" . 
 			"SEGMENTATION_COMMAND = A string (enclosed by quotes) that specifies the segmenter to run" .
@@ -34,10 +34,12 @@ my $usage = "\nK-Fold Cross-Validater\n\n" .
 			"\t-e FLAGS\tFlags to pass on to errors.pl. (Must enclose in quotes.)\n" .
 			"\t-k NUM_FOLDS\tSpecifies number of folds for cross-validation. (Default = 10)\n" .
 			"\t-p PREFIX\tFile prefix for temporary fold files. Note: All files with this prefix are DELETED when script ends. (Default = 'cvfoldtemp')\n" .
-			"\t-d \t\tDo NOT delete files that begin with PREFIX when done. (Although any existing ones are still destroyed when the script is first run.)\n";
+			"\t-d \t\tDo NOT delete files that begin with PREFIX when done. (Although any existing ones are still destroyed when the script is first run.)\n" .
+			"\t-i \t\tInvert training and test sets.  With 10 folds, that would mean train on 1 and test on 9.\n" .
+			"\t-v \t\tOutput result of running errors.pl for each individual fold.\n";
 
 # Handle arguments
-getopts('e:k:p:d');
+getopts('e:k:p:div');
 die $usage if @ARGV < 2;
 my $segmenter = shift @ARGV;
 my $corpus = shift @ARGV;
@@ -77,22 +79,55 @@ for (my $i = 0; $i < $folds; $i++)
 	{
 		`rm $prefix-gold-current`;		
 	}
-	splice(@currentFiles, $i, 1);
-	for (my $j = 0; $j < $folds - 1; $j++)
+	if ($opt_i)
 	{
-		$ir = rand @currentFiles;
-		push @selection, $currentFiles[$ir];
-		splice(@currentFiles, $ir, 1);
+		`cat $allFiles[$i] >> $prefix-gold-current`;
+		splice(@currentFiles, $i, 1);
+		for (my $j = 0; $j < $folds - 1; $j++)
+		{
+			$ir = rand @currentFiles;
+			push @selection, $currentFiles[$ir];
+			splice(@currentFiles, $ir, 1);
+		}
+		foreach my $f (@selection)
+		{
+			`cat $f >> $prefix-gold-current`;
+		}
+		print "Segmenting fold $i...";
+		`$segmenter $foldSize $prefix-gold-current | tail -n $trainingSize >> $prefix-results`;
+		`tail -n $trainingSize $prefix-gold-current >> $prefix-gold`;
 	}
-	foreach my $f (@selection)
+	else
 	{
-		`cat $f >> $prefix-gold-current`;
+		splice(@currentFiles, $i, 1);
+		for (my $j = 0; $j < $folds - 1; $j++)
+		{
+			$ir = rand @currentFiles;
+			push @selection, $currentFiles[$ir];
+			splice(@currentFiles, $ir, 1);
+		}
+		foreach my $f (@selection)
+		{
+			`cat $f >> $prefix-gold-current`;
+		}
+		`cat $allFiles[$i] >> $prefix-gold-current`;		
+		print "Segmenting fold $i...";
+		`$segmenter $trainingSize $prefix-gold-current | tail -n $foldSize >> $prefix-results`;
+		`tail -n $foldSize $prefix-gold-current >> $prefix-gold`;
 	}
-	`cat $allFiles[$i] >> $prefix-gold-current`;
-	print "Segmenting fold $i...";
-	`$segmenter $trainingSize $prefix-gold-current | tail -n $foldSize >> $prefix-results`;
-	`tail -n $foldSize $prefix-gold-current >> $prefix-gold`;
 	print "done\n";
+	if ($opt_v)
+	{
+		print `./errors.pl $opt_e $prefix-gold $prefix-results`;
+		`rm $prefix-gold`;
+		`rm $prefix-results`;
+	}
 }
-print `./errors.pl $opt_e $prefix-gold $prefix-results`;
-`rm $prefix*`;
+if (!$opt_v)
+{
+	print `./errors.pl $opt_e $prefix-gold $prefix-results`;			
+}
+if (!$opt_d)
+{
+	`rm $prefix*`;	
+}
